okay nicer way of moving stuff around
