+++ maitake/src/time.rs
@@ -1,6 +1,4 @@
 //! Time utilities.
-mod sleep;
 mod timer;
 
-pub use self::sleep::Sleep;
-pub use self::timer::Timer;
+pub use self::timer::{Sleep, Timer};
